Fix type in test

+++ nano/core_test/difficulty.cpp
@@ -26,7 +26,7 @@ TEST (difficulty, multipliers)
 	{
 		uint64_t base = 0xffffffc000000000;
 		uint64_t difficulty_nil = 0;
-		uint64_t multiplier_nil = 0.;
+		double multiplier_nil = 0.;
 #ifndef NDEBUG
 		ASSERT_DEATH_IF_SUPPORTED (nano::difficulty::to_multiplier (difficulty_nil, base), "");
 		ASSERT_DEATH_IF_SUPPORTED (nano::difficulty::from_multiplier (multiplier_nil, base), "");
